엔딩 및 잡다한거

+++ Assets/@Scripts/Data/Datas.cs
@@ -106,7 +106,7 @@ namespace Data
         public int WaveIdx;
         public int WaveNumber;
         public int WaveTimeLimit;
-        public int MobSpawnRate;
+        public float MobSpawnRate;
         public int NextWaveIdx;
         public List<int> MobIDs;
